requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
plotly>=5.15.0
streamlit>=1.25.0

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
akshare>=1.12.0
openpyxl>=3.1.0

pdfplumber>=0.11.0
python-dotenv>=1.0.0
dashscope>=1.14.0
httpx>=0.25.0

reportlab>=4.0.0

# FastAPI backend
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
